+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+0.4.1
+---
+* Added `filter()` method to Note and Email to let you pull all notes/emails from a deal, case, person, or company
+* Added `emails` and `notes` properties to Deal, Person, and Company objects as a shortcut to get all their notes/emails
+
+
 0.4.0
 ---
 * Added support for Emails